Move handling of /r command-line parameter to OnEventLoopEnter
Since the event loop doesn't exist yet when OnCmdLineParsed is called,
the profiling progress dialog is unresponsive to input events. This
made it impossible for profiling to stop in any way other than the
program terminating.

Options were also not loaded at that point; thus, the SpeedThrottle
option was being ignored when using /r.