Lower profiler thread priority when sampling ends
The sampling is frequently interrupted with Sleep calls, which allows
the GUI to remain responsive. However, the "symbol query" phase is very
CPU intensive. Running it at TIME_CRITICAL priority causes the GUI,
and seemingly even the system as a whole, to become unresponsive.

Since critical priority isn't of much use once the sampling is done,
lower the profiler thread priority to NORMAL.

I took the liberty to do a bit of refactoring in MyThread for the sake of
encapsulating operations on the thread handle.