Build wxWidgets with Very Sleepy
This greatly simplifies building wxWidgets for Very Sleepy, as it no
longer requires manually downloading/extracting wxWidgets, invoking
nmake for each platform/configuration pair, and setting up an
environment variable that is visible to Visual Studio which indicates
the wxWidgets directory.

Unfortunately, this comes with a drawback: the wxWidgets Visual Studio
project files only support building against the DLL version of the C
runtime, which means that we are forced to switch away from statically
linking Very Sleepy to the C runtime, to linking dynamically to the
DLL version, and requiring that the DLL is present on the system Very
Sleepy runs on.

Installer changes to facilitate the above precondition will follow.