Show thread name on threads list, if available. (#60)
Find if GetThreadDescription() is available in Kernel32.dll (it's
available from the Win10 Creators update onwards). If it is, add
another column "Thread Name" to the list of threads shown when
selecting which thread(s) to profile. If this function is unavailable,
don't show this extra column at all. If it is available, we rely on
the application being profiled to have called SetThreadDescription(),
but if it has not, we just show "-" instead.