Fix bad sample scaling when pausing
The first sample after resuming after a pause was disproportionately
scaled to the entire duration of the pause.

The simplest correct way to approach pausing is to execute the entire
loop always, except skip the actual sampling when we're paused.  The
main downside to this approach is that we will spin the CPU even when
paused, if the sampling interval is 0.

Instead, fake the timestamp of the previous sample to the current time
when exiting a pause. This effectively discards the first sample after
resuming a pause (since we don't have an accurate sample duration for
it anyway).